+++ database/upgrade/upgrade3.11.0.sql
@@ -68,6 +68,7 @@ DROP TABLE myems_system_db.tbl_microgrid_architecture_types;
 DROP TABLE myems_system_db.tbl_microgrid_owner_types;
 ALTER TABLE myems_system_db.tbl_microgrids DROP COLUMN architecture_type_id;
 ALTER TABLE myems_system_db.tbl_microgrids DROP COLUMN owner_type_id;
+ALTER TABLE myems_system_db.tbl_microgrids ADD `serial_number` VARCHAR(255) AFTER `cost_center_id`;
 DROP TABLE myems_system_db.tbl_microgrids_windturbines;
 
 -- UPDATE VERSION NUMBER