+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added energy storage power station to database, myems-api, and myems-admin
 - added photovoltaic power station to database, myems-api, and myems-admin
 - added wind farm to database, myems-api, and myems-admin
-- 
+- added serial number to microgrid
 ### Changed
 - changed dropdown menu columns from 3 to 2 in myems-web
 -